Set time in Homescreen and send to Timer
